# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API completa para gestão de equinos com Tokenização RWA (Real World Assets). Inclui: autenticação, gestão de equinos, tokenização, leilões, exames laboratoriais, nutrição com IA, treinamento, rankings, financeiro e muito mais.",
        "title": "Equinoid API - Sistema Perfeito",
        "termsOfService": "https://equinoid.com/terms",
        "contact": {
            "name": "Equinoid Support",
            "url": "https://equinoid.com/support",
            "email": "suporte@equinoid.com"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://equinoid.com/license"
        },
        "version": "2.0"
    },
    "host": "seahorse-app-28du8.ondigitalocean.app",
    "basePath": "/api/v1",
    "paths": {
...
        "/exames-laboratoriais": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os exames laboratoriais com filtros opcionais",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Listar exames laboratoriais",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por Equinoid",
                        "name": "equinoid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por tipo de exame",
                        "name": "tipo_exame",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por veterinário",
                        "name": "veterinario_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por laboratório",
                        "name": "laboratorio_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova solicitação de exame laboratorial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Criar solicitação de exame",
                "parameters": [
                    {
                        "description": "Dados do exame",
                        "name": "exame",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateExameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exames-laboratoriais/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados completos de um exame laboratorial",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Buscar exame por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do exame",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza informações de um exame laboratorial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Atualizar exame",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do exame",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "exame",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateExameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um exame laboratorial do sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Deletar exame",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do exame",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exames-laboratoriais/{id}/concluir": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finaliza o exame registrando resultado, valores e laudo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Concluir exame com resultado",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do exame",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resultado do exame",
                        "name": "resultado",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "laudo": {
                                    "type": "string"
                                },
                                "resultado": {
                                    "type": "string"
                                },
                                "valores": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exames-laboratoriais/{id}/iniciar-analise": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marca o início da análise laboratorial",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Iniciar análise do exame",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do exame",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exames-laboratoriais/{id}/receber-amostra": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra o recebimento da amostra no laboratório",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exames"
                ],
                "summary": "Receber amostra no laboratório",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do exame",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data de recebimento",
                        "name": "recebimento",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data_recebimento": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateExameRequest": {
            "type": "object",
            "required": [
                "equinoid",
                "nome_exame",
                "tipo_exame",
                "veterinario_solicitante_id"
            ],
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "equinoid": {
                    "type": "string"
                },
                "laboratorio_id": {
                    "type": "integer"
                },
                "nome_exame": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "tipo_exame": {
                    "type": "string"
                },
                "veterinario_solicitante_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.ResultadoExame": {
            "type": "string",
            "enum": [
                "normal",
                "alterado",
                "positivo",
                "negativo",
                "inconclusivo"
            ],
            "x-enum-varnames": [
                "ResultadoNormal",
                "ResultadoAlterado",
                "ResultadoPositivo",
                "ResultadoNegativo",
                "ResultadoInconclusivo"
            ]
        },
...
        "models.UpdateExameRequest": {
            "type": "object",
            "properties": {
                "data_coleta": {
                    "type": "string"
                },
                "data_conclusao": {
                    "type": "string"
                },
                "data_inicio_analise": {
                    "type": "string"
                },
                "data_recebimento_amostra": {
                    "type": "string"
                },
                "laudo": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "resultado": {
                    "$ref": "#/definitions/models.ResultadoExame"
                },
                "status": {
                    "type": "string"
                },
                "valores": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}